↺ merge  svn merge -r 1602:1680 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/branches/yui/WORK_01
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@1681 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ support_gimic/codecnv/eucsjis.c
@@ -39,7 +39,7 @@ UINT codecnv_euctosjis(char *lpOutput, UINT cchOutput, const char *lpInput, UINT
 	else
 	{
 		// String mode
-		nLength = euctosjis(lpOutput, cchOutput - 1, lpInput, strlen(lpInput));
+		nLength = euctosjis(lpOutput, cchOutput - 1, lpInput, (UINT)strlen(lpInput));
 		if (lpOutput)
 		{
 			lpOutput[nLength] = '\0';
